swa-deploy/api/requirements.txt
azure-cosmos==4.7.0
python-dateutil==2.9.0
